+++ packages/faas/CHANGELOG.md
@@ -6,6 +6,16 @@
 
 - Breaking change: removed `uid` in script model, now replaced with `scopeId`.
 
+- Breaking change: lambda KV storage now must need a `mongodb database engine`.
+
 - Feat: added support for `@tigojs/faas-log` (due to the breaking changes, the addon will not work with older version).
+
+- Feat: added a method called `getName` to get lambda name by ID.
+
+- Fix: the default `Content-Type` header will be `text/plain` by default when `body` in `respondWith` method is not an object.
+
+- Fix: lambda KV storage key and cache key is not right.
+
+- Fix: service can read lambda kv config correctly now.
 
 - Minor: upgrade `@tigojs/utils`.